Add radionuclide edep contributions
Add the ability for the radionuclide source to contribute to the dose
within the source region. This required some additions to the
application class and dose scoring object, to allow a region to be
passed as an argument. The region of the top particle on the stack could
not be used because there may not be a particle on the stack during the
first call of getNextParticle.

The dose scoring for the radionuclide source was necessary to include
sub-threshold contributions from relaxations, and local absorption of
alpha emissions.