Fix #189: Fix bug in EDKnrc when using photon spectrum with source 0
Instead of calling shower, the program uses macro $DO_PHOTON_SHOWER
which failed to set the value of GMFPR0 for the initial interaction
at the origin.  The macro also did not handle cases where intial
energy, EI, was <= PCUT, instead allowing GMFPR0 to be set to 0.  The
result of both bugs was many "Infinite" values.  These
bugs have been fixed.  Also fixed a warning in geomsph.mortran in which
there was an incompatibility in the variable passing TITLE to
xvgrplot.